Split RelSeries.jl.
